v1.0
this is the first fully working version. it allows the playback of audio along with making the sections from the time stamps in that audio. there's not really any error handling or checks against user error, that may or may not come later. slider still needs to change the song time.

+++ BeatSaver1/BeatSaverEntity.cs
@@ -9,10 +9,10 @@ namespace BeatSaver1
     public class BeatSaverEntity
     {
         public string _version { get; set; }
-        public int _beatsPerMinute { get; set; }
-        public int _beatsPerBar { get; set; }
-        public int _noteJumpSpeed { get; set; }
-        public int _shuffle { get; set; }
+        public decimal _beatsPerMinute { get; set; }
+        public decimal _beatsPerBar { get; set; }
+        public decimal _noteJumpSpeed { get; set; }
+        public decimal _shuffle { get; set; }
         public decimal _shufflePeriod { get; set; }
         public EventsEntity[] _events { get; set; }
         public NotesEntity[] _notes { get; set; }
